requirements.txt
click
querystring_parser
grpcio
grpcio-tools
google-api
google-api-core
protoc_gen_swagger
aiohttp
sqlalchemy
databases[postgresql]
pymongo
mock
pytest
black
flake8
flake8-bugbear
flake8-builtins
flake8-blind-except
mypy